Bug fixes, added missing functions, and resolved module import issues.
