+++ MistIllum.js
@@ -1,12 +1,106 @@
 import nvk from 'nvk';
-import x11 from 'node-x11';
+import x11 from 'x11';
 import { exec } from 'node:child_process';
 import clipboardy from 'clipboardy';
 import { renderViewport, selectTimeIndex, selectCategory, selectItem } from './MistCore.js';
 import { SelectionModeState, MetricTensorND } from './MistCommon.js';
-import { startSession, loadMistUser, milestoneManager, handleSelectionBackend } from './MistTrackerVulkan.js';
+import { startSession, loadMistUser, MilestoneManager } from './MistTrackerVulkan.js';
 import { mistSolution } from './mistSolution.js';
-import * as MistMulti from './MistMulti.js';
+import * as MistMulti from './MistMulti.cjs';
+
+/**
+ * Calculate the probability of an event occurring based on quantum mechanical principles
+ * and wave function interference patterns.
+ * 
+ * This implementation considers:
+ * 1. Quantum state of involved objects
+ * 2. Wave function interference
+ * 3. Energy distribution across dimensions
+ * 4. Temporal causality
+ * 5. Object locality and entanglement
+ * 
+ * @param {Object} event - The event to evaluate
+ * @param {Object} event.objects - Array of objects involved in the event
+ * @param {Array} event.timeVector - [quantum time, interaction time, cosmological time]
+ * @param {Object} event.energy - Energy distribution for the event
+ * @param {Array} event.position - Position vector in nD space
+ * @returns {number} Probability between 0 and 1
+ */
+export function probabilityOfEvent(event) {
+    if (!event || !event.objects || !event.timeVector) {
+        return 0; // Invalid event
+    }
+
+    // 1. Calculate wave function contributions
+    const waveFunctions = event.objects.map(obj => {
+        return obj.waveFunction || waveFunction(1, 1, 0, 1, event.timeVector);
+    });
+
+    // 2. Calculate interference patterns between objects
+    let interference = 1;
+    for (let i = 0; i < event.objects.length; i++) {
+        for (let j = i + 1; j < event.objects.length; j++) {
+            interference *= interferencePattern(
+                event.objects[i].position || [0, 0, 0],
+                event.objects[i].position || [0, 0, 0],
+                event.objects[j].position || [0, 0, 0],
+                event.energy?.wavelength || 1
+            );
+        }
+    }
+
+    // 3. Energy distribution factor
+    const energyDistribution = event.energy ? distributeEnergy(
+        event.energy.total || 1,
+        ['quantum', 'interaction', 'cosmological', 'spatial']
+    ) : { quantum: 0.25, interaction: 0.25, cosmological: 0.25, spatial: 0.25 };
+
+    // 4. Locality check
+    const localityFactor = event.objects.reduce((factor, obj1, i) => {
+        return factor * event.objects.slice(i + 1).reduce((f, obj2) => {
+            return f * locality(
+                obj1.position || [0, 0, 0],
+                obj2.position || [0, 0, 0]
+            );
+        }, 1);
+    }, 1);
+
+    // 5. Bell's theorem test for quantum correlations
+    const bellFactor = bellTheorem(
+        event.objects[0]?.quantumState?.[0] || 0,
+        event.objects[0]?.quantumState?.[1] || 0,
+        event.objects[1]?.quantumState?.[0] || 0,
+        event.objects[1]?.quantumState?.[1] || 0
+    );
+
+    // 6. Pilot wave contribution
+    const pilotWaveFactor = event.objects.reduce((factor, obj) => {
+        const pwave = pilotWave(
+            obj.waveFunction || waveFunction(1, 1, 0, 1, event.timeVector),
+            { energy: energyDistribution }
+        );
+        return factor * (typeof pwave === 'number' ? pwave : 1);
+    }, 1);
+
+    // 7. Temporal causality check
+    const [quantumTime, interactionTime, cosmologicalTime] = event.timeVector;
+    const temporalFactor = Math.exp(
+        -(Math.abs(quantumTime) + Math.abs(interactionTime) + Math.abs(cosmologicalTime))
+    );
+
+    // Combine all factors
+    const rawProbability = (
+        Math.abs(interference) *
+        energyDistribution.quantum *
+        localityFactor *
+        Math.abs(bellFactor) *
+        pilotWaveFactor *
+        temporalFactor
+    );
+
+    // Ensure return value is between 0 and 1
+    return Math.max(0, Math.min(1, rawProbability));
+}
 
 class MistIllum {
     constructor(config = {}) {
@@ -1179,7 +1273,7 @@ function launchMistCore(config = {}) {
  * @param {Object} uiRenderer - UI rendering interface.
  */
 function launchMistMulti(config = {}, uiRenderer = global.uiRenderer) {
-  const MistMulti = require('./MistMulti.js');
+  const MistMulti = require('./MistMulti.cjs');
   const db = config.db || null;
   const menuControl = new MistMenuControl(db, uiRenderer);
   menuControl.start(config.userName || 'guest');
@@ -1945,7 +2039,7 @@ function getMenuOptionsWithMilestones() {
 
 // --- Export API ---
 
-module.exports = {
+export {
   tileMode,
   tileSpan,
   LightSource,
